requirements.txt
gradio
pdfminer.six
huggingface_hub
pymupdf